move booking actions into a subfolder
